skip reading object kind when reading array items

+++ src/buffer.c
@@ -36,7 +36,7 @@ t_ion_result_code ion_buffer_seek(t_ion_buffer *self, uint8_t curr) {
   if (self == NULL)
     return RESULT_ERROR;
 
-  return vector_seek(self->body, curr);
+  return vector_seek(self->body, (size_t) curr);
 }
 
 t_ion_result_code ion_buffer_write(t_ion_buffer *self, void *src, size_t len) {
